completed model server integration + normalized to UUID instead of str in backend

+++ backend/app/utils/storage_helper.py
@@ -1,3 +1,4 @@
+import logging
 from typing import Optional
 from uuid import UUID
 
@@ -6,12 +7,14 @@ from supabase import AsyncClient
 from app.core.db_dependencies import DBPool
 from app.database import notes_db
 
+logger = logging.getLogger(__name__)
+
 
 async def cleanup_storage(
-    user_id: str,
+    user_id: UUID,
     db: DBPool,
     supabase_client: AsyncClient,
-    session_id: Optional[str] = None,
+    session_id: Optional[UUID] = None,
 ) -> None:
     """
     Takes in user id and session id and deletes all files in the corresponding storage path in supabase.
@@ -19,14 +22,17 @@ async def cleanup_storage(
     """
     bucket = "session-pdfs"
 
-    path = f"{user_id}/{session_id}"
+    path = f"{user_id}/{session_id}" if session_id else f"{user_id}"
     files = await supabase_client.storage.from_(bucket).list(path)
     file_paths = [f"{path}/{file['name']}" for file in files if file.get("name")]
     if file_paths:
         try:
             await supabase_client.storage.from_(bucket).remove(file_paths)
-            await notes_db.delete_all_notes(
-                db=db, session_id=UUID(session_id), user_id=UUID(user_id)
-            )
+            if session_id:
+                await notes_db.delete_all_notes(
+                    db=db,
+                    session_id=session_id,
+                    user_id=user_id,
+                )
         except Exception as e:
-            print(f"Error deleting files from storage: {e}")
+            logger.exception("Error deleting files from storage: %s", e)
